refactor sifter parsing into its own module
